+++ src/qt/widgets/floatsliderwidgetqt.cpp
@@ -103,7 +103,7 @@ void FloatSliderWidgetQt::setSpinBoxDecimals( float increment ) {
 
     std::string str2 = buff.str().substr(buff.str().find(".")+1,buff.str().length());
 
-    spinBox_->setDecimals(str2.length());
+    spinBox_->setDecimals(static_cast<int>(str2.length()));
 
 }
